requirements.txt
flask==3.0.3
python-dotenv==1.0.1
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9

langchain==0.2.16
langgraph==0.2.35

pydantic==2.8.2
python-dateutil==2.9.0.post0

amadeus==12.0.0

